feat: 일정 생성 기능 추가

+++ BE/src/main/java/com/codaily/management/dto/CalendarResponse.java
@@ -28,7 +28,7 @@ public class CalendarResponse {
         private String featureTitle;
         private String featureDescription;
         private LocalDate scheduleDate;
-        private Integer allocatedHours;
+        private Double allocatedHours;
         private String category;
         private Integer priorityLevel;
         private String status;
